app logic fixed

+++ application/archive.py
@@ -7,10 +7,14 @@ import zipfile
 
 def create_archive(path):
     arc_name = f"{path[2:]}.zip"
-    if os.path.exists(path):
-        with zipfile.ZipFile(arc_name, 'w', zipfile.ZIP_DEFLATED) as zipf:
-            for image in os.listdir(path):
-                image_path = os.path.join(path, image)
-                zipf.write(image_path)
-        shutil.rmtree(path)
-        return arc_name
+    if not os.path.exists(arc_name):
+        if os.path.exists(path):
+            with zipfile.ZipFile(arc_name, 'w', zipfile.ZIP_DEFLATED) as zipf:
+                for image in os.listdir(path):
+                    image_path = os.path.join(path, image)
+                    zipf.write(image_path)
+            shutil.rmtree(path)
+        else:
+            return
+
+    return arc_name
